[firebase] Add expo-firebase-app on iOS
The expo-firebase-app package manages the lifecycle of the firebase-apps. It contains the following functionality:
- auto initialize firebase app based in case a GoogleService-Info.plist exists
- auto update firebase-app when Expo client loads a new bundle/manifest
- methods to create/delete/read firebase apps from JS

The idea is to let this package orchestrate all firebase-app creation and deletion. Whenever an app is reloaded, this module checks whether the default firebase app matches the configured GoogleService-Info.plist. If an app was loaded that does not have a GoogleService-Info file, then the Firebase app will be deleted.

Not yet working:
- support multiple firebase apps